удаляет лишние файлы
