--- js/components/handlers/back-button-handler.js
@@ -1,11 +1,0 @@
-import changeView from '../../router/change-view';
-
-/**
- * Возвращает на экран приветствия
- * @param {MouseEvent} evt
- */
-const onBackButtonClicked = () => {
-  changeView(`greeting`);
-};
-
-export default onBackButtonClicked;
