+++ js/data/state.js
@@ -6,38 +6,55 @@ const AnswersMap = {
   FALSE: `incorrect`
 };
 
-export class InitialState {
-  constructor() {
-    this.lives = 3;
-    this.currentScreen = ``;
-    this.timeLeft = 90;
-    this.playerName = ``;
-    this.answers = [];
-  }
 
-  /**
-   * Сохраняет имя пользователя
-   * @param {string} name
-   */
-  setName(name) {
-    this.playerName = name;
+const getType = (timeLeft) => {
+  if (30 - timeLeft < 10) {
+    return `fast`;
+  } else if (30 - timeLeft > 20) {
+    return `slow`;
   }
+  return `normal`;
+};
 
-  /**
-   * Добавляет ответ
-   * @param {string} correctness
-   * @param {string} type
-   */
-  addAnswer(correctness, type) {
-    this.answers.push({
-      correctness: AnswersMap[correctness.toUpperCase()],
-      type
-    });
-  }
+const state = {
+  lives: 3,
+  timeLeft: 30,
+  playerName: ``,
+  answers: []
+};
 
-  decreaseLives() {
-    this.lives--;
-  }
-}
+state.setName = (oldState, name) => {
+  const newState = Object.assign({}, oldState);
+  newState.playerName = name;
+  return newState;
+};
 
-export const state = new InitialState();
+state.setTime = (oldState, time) => {
+  const newState = Object.assign({}, oldState);
+  newState.timeLeft = time;
+  return newState;
+};
+
+state.resetTime = (oldState) => {
+  const newState = Object.assign({}, oldState);
+  newState.timeLeft = 30;
+  return newState;
+};
+
+state.setLives = (oldState, lives) => {
+  const newState = Object.assign({}, oldState);
+  newState.lives = lives;
+  return newState;
+};
+
+state.addAnswer = (oldState, correctness) => {
+  const newState = Object.assign({}, oldState);
+  newState.answers.push({
+    correctness: AnswersMap[correctness.toUpperCase()],
+    type: AnswersMap[correctness.toUpperCase()] === `correct` ? getType(oldState.timeLeft) : `normal`
+  });
+  return newState;
+};
+
+export default state;
+
