+++ js/main.js
@@ -1,6 +1,3 @@
-// import changeView from './router/change-view';
-// import state from './data/state2';
-
 import App from './application';
 
 App.showIntro();
